sd task #645 review rework reply and reply_bytes logic exchange tests fix fill mailbox with empty vector if given program id doesn't exist
