+++ gtest/src/program.rs
@@ -251,15 +251,8 @@ pub fn calculate_program_id(code_hash: CodeId, salt: &[u8]) -> ProgramId {
 
 #[cfg(test)]
 mod tests {
-    use codec::Encode;
-    use gear_core::{
-        ids::MessageId,
-        message::{Dispatch, DispatchKind, Message, Payload},
-    };
-
+    use super::Program;
     use crate::{Log, System};
-
-    use super::{Program, ProgramIdWrapper};
 
     #[test]
     fn test_handle_messages_to_failing_program() {
@@ -283,148 +276,5 @@ mod tests {
 
         assert!(!run_result.main_failed());
         assert!(run_result.contains(&expected_log));
-    }
-
-    #[test]
-    fn mailbox_mock_walkthrough_test() {
-        let system = System::new();
-        let message_id: MessageId = Default::default();
-        let source_user_id = ProgramIdWrapper::from(100).0;
-        let destination_user_id = ProgramIdWrapper::from(200).0;
-        let message_payload: Payload = vec![1, 2, 3].into();
-        let encoded_message_payload: Payload = message_payload.encode().into();
-        let reply_payload: Payload = vec![3, 2, 1].into();
-        let log = Log::builder().payload(message_payload.clone());
-        let encoded_reply_payload: Payload = reply_payload.encode().into();
-
-        let message = Message::new(
-            message_id,
-            source_user_id,
-            destination_user_id,
-            encoded_message_payload.clone(),
-            Default::default(),
-            2,
-            None,
-        );
-
-        let message_result =
-            system.send_dispatch(Dispatch::new(DispatchKind::Handle, message.clone()));
-        let message_log = message_result
-            .log
-            .last()
-            .expect("No message log in run result");
-
-        let destination_user_mailbox = system.get_mailbox(destination_user_id);
-        let message_replier = destination_user_mailbox.take_message(log);
-        let reply_log = message_replier.reply(reply_payload.clone(), 1).log;
-
-        let last_reply_log = reply_log.last().expect("No message log in run result");
-
-        let second_message_result =
-            system.send_dispatch(Dispatch::new(DispatchKind::Handle, message));
-
-        let second_message_log = message_result
-            .log
-            .last()
-            .expect("No message log in run result");
-
-        assert!(!message_result.main_failed);
-        assert!(!message_result.others_failed);
-        assert!(!second_message_result.main_failed);
-        assert!(!second_message_result.others_failed);
-        assert_eq!(reply_log.len(), 1);
-        assert_eq!(last_reply_log.payload(), encoded_reply_payload);
-        assert_eq!(message_log.payload(), encoded_message_payload);
-        assert_eq!(second_message_log.payload(), encoded_message_payload);
-    }
-
-    #[test]
-    fn mailbox_mock_deletes_message_after_reply() {
-        let system = System::new();
-        let message_id: MessageId = Default::default();
-        let source_user_id = ProgramIdWrapper::from(100).0;
-        let destination_user_id = ProgramIdWrapper::from(200).0;
-        let message_payload: Payload = vec![1, 2, 3].into();
-        let reply_payload: Payload = vec![3, 2, 1].into();
-        let message_log = Log::builder().payload(message_payload.clone());
-
-        let message = Message::new(
-            message_id,
-            source_user_id,
-            destination_user_id,
-            message_payload.encode().into(),
-            Default::default(),
-            2,
-            None,
-        );
-
-        system.send_dispatch(Dispatch::new(DispatchKind::Handle, message));
-
-        let mut destination_user_mailbox = system.get_mailbox(destination_user_id);
-        let message_replier = destination_user_mailbox.take_message(message_log.clone());
-        message_replier.reply(reply_payload, 1);
-
-        destination_user_mailbox = system.get_mailbox(destination_user_id);
-        assert!(!destination_user_mailbox.contains(message_log))
-    }
-
-    #[test]
-    fn mailbox_mock_reply_bytes_test() {
-        let system = System::new();
-        let message_id: MessageId = Default::default();
-        let source_user_id = ProgramIdWrapper::from(100).0;
-        let destination_user_id = ProgramIdWrapper::from(200).0;
-        let message_payload: Payload = vec![1, 2, 3].into();
-        let reply_payload_array: [u8; 3] = [3, 2, 1];
-        let reply_payload: Payload = reply_payload_array.to_vec().into();
-        let encoded_reply_payload: Payload = reply_payload.encode().into();
-        let log = Log::builder().payload(message_payload.clone());
-
-        let message = Message::new(
-            message_id,
-            source_user_id,
-            destination_user_id,
-            message_payload.encode().into(),
-            Default::default(),
-            2,
-            None,
-        );
-
-        system.send_dispatch(Dispatch::new(DispatchKind::Handle, message));
-
-        let destination_user_mailbox = system.get_mailbox(destination_user_id);
-        let message_replier = destination_user_mailbox.take_message(log);
-
-        let result = message_replier.reply_bytes(&reply_payload_array, 1);
-        let result_log = result.log;
-        let last_result_log = result_log.last().expect("No message log in run result");
-        assert_eq!(last_result_log.payload(), encoded_reply_payload);
-    }
-
-    #[test]
-    fn mailbox_mock_deletes_message_after_taking() {
-        let system = System::new();
-        let message_id: MessageId = Default::default();
-        let source_user_id = ProgramIdWrapper::from(100).0;
-        let destination_user_id = ProgramIdWrapper::from(200).0;
-        let message_payload: Payload = vec![1, 2, 3].into();
-        let log = Log::builder().payload(message_payload.clone());
-
-        let message = Message::new(
-            message_id,
-            source_user_id,
-            destination_user_id,
-            message_payload.encode().into(),
-            Default::default(),
-            2,
-            None,
-        );
-
-        system.send_dispatch(Dispatch::new(DispatchKind::Handle, message));
-
-        let destination_user_mailbox = system.get_mailbox(destination_user_id);
-        destination_user_mailbox.take_message(log.clone());
-
-        assert!(!destination_user_mailbox.contains(log))
     }
 }
